mcp/python-google-adk-trusted-auth/requirements.txt
google-adk==1.12.0
python-dotenv>=1.0.0
google-genai>=0.3.0
aiohttp>=3.8.0
rich>=13.0.0
Flask>=2.2.5
requests>=2.32.4
